simQt: Low Hanging Qt6 Updates
**JIRA Issue:** SIM-17935

**Description:** Various low-hanging fixes for compatibility with Qt6. Includes various changes including:
* Improved (but not yet complete) Qt5/Qt6 handling in CMake, using version-agnostic variables/macros where needed
* Added missing includes
* New SIMQT_QT_VERSION value based on detection of Qt5/6. Updated simQtConfig.cmake to respect correct version
* Removed setMargin() (deprecated) in favor of setContentsMargins()
* Removed unused deprecated headers
* Fixed wide string problem in Font
* Conditionally compile QGLFormat for Gl3FormatGuesser
* Removed no-longer-needed `qRegisterMetaTypeStreamOperators`
* Fixed template errors due to type changes in Qt
* enterEvent parameter fixes for Toast, with override keywords to get compiler to help catch the issue

**Notes:** First of many passes for Qt6 support.

**Testing Performed:** These files compile with Qt6 support, with other local changes to build against Qt6. All of these were found due to compiler errors during a Qt6 compile. Tested the margins explicitly on Qt5 builds, and checked output of simQtConfig.cmake on both Qt5 and Qt6 builds and it looks right in both.

+++ SDK/simQt/SegmentedSpinBox.cpp
@@ -286,13 +286,13 @@ private:
     const int selectionLength = lineEdit()->selectedText().length();
     const int cursorPosition = lineEdit()->cursorPosition();
     lineEdit()->setText(completeLine_->text());
-    lineEdit()->setCursorPosition(simCore::sdkMin(cursorPosition, lineEdit()->text().length()));
+    lineEdit()->setCursorPosition(simCore::sdkMin(cursorPosition, static_cast<int>(lineEdit()->text().length())));
     // If there was a selection, restore it after timestamp is updated
     if (selectionStart != -1)
     {
-      const int newSelectionStart = simCore::sdkMin(selectionStart, lineEdit()->text().length() - 1);
+      const int newSelectionStart = simCore::sdkMin(selectionStart, static_cast<int>(lineEdit()->text().length()) - 1);
       if (newSelectionStart != -1)
-        lineEdit()->setSelection(newSelectionStart, simCore::sdkMin(selectionLength, lineEdit()->text().length() - newSelectionStart));
+        lineEdit()->setSelection(newSelectionStart, simCore::sdkMin(selectionLength, static_cast<int>(lineEdit()->text().length()) - newSelectionStart));
     }
 
     if (lastEditedTime_ != completeLine_->timeStamp())
